requirements.txt
streamlit
faiss-cpu
sentence-transformers
numpy
requests
google-generativeai
streamlit-extras
pickle5
